feat: convert login system into developer plugin
