+++ README.md
@@ -243,9 +243,12 @@ To enable optional Google authentication, set `google_client_id` in
 `settings.json` or provide it via the environment variable
 `DOCROPPER_GOOGLE_CLIENT_ID`. The login module is active only when
 `license_check` is enabled; otherwise the sign-in button remains hidden even if
-a client ID is provided. When configured, the web interface displays the
-button and tokens are verified by the backend. When the hidden Demo Full license
-is active the login button is hidden even if `google_client_id` is set.
+a client ID is provided. The module is implemented as a plugin and currently
+flagged developer-only (`DOCROPPER_LOGIN_DEV_ONLY` / `login_dev_only`), so it
+loads only for developer licenses until finished. When configured, the web
+interface displays the button and tokens are verified by the backend. When the
+hidden Demo Full license is active the login button is hidden even if
+`google_client_id` is set.
 
 ---
